\name{disdat-package}

\alias{disdat-package}
\alias{disdat}

\docType{package}

\title{
Species distribution modeling
}

\description{
This package has datasets for comparing species distribution models. These are the AWT data that were used in this model comparison study: 

Elith, J., C.H. Graham, R.P. Anderson, M. Dudik, S. Ferrier, A. Guisan, R.J. Hijmans, F. Huettmann, J. Leathwick, A. Lehmann, J. Li, L.G. Lohmann, B. Loiselle, G. Manion, C. Moritz, M. Nakamura, Y. Nakazawa, J. McC. Overton, A.T. Peterson, S. Phillips, K. Richardson, R. Scachetti-Pereira, R. Schapire, J. Soberon, S. Williams, M. Wisz and N. Zimmerman, 2006. Novel methods improve prediction of species' distributions from occurrence data. Ecography 29: 129-151. \url{http://dx.doi.org/10.1111/j.2006.0906-7590.04596.x}

}

\details{
\tabular{ll}{
Package: \tab disdat\cr
Type: \tab Package\cr
Version: \tab 0.0\cr
Date: \tab 28-March-2010\cr
License: \tab GPL 3\cr
LazyLoad: \tab yes\cr
}
}

\author{
Robert J. Hijmans, Jane Elith and Steven Phillips.
}

\keyword{ package }
\keyword{ spatial }

