\name{disdat-package}

\alias{disdat-package}
\alias{disdat}

\docType{package}

\title{
Species distribution modeling
}

\description{
This package has datasets for comparing species distribution models.
}

\details{
\tabular{ll}{
Package: \tab disdat\cr
Type: \tab Package\cr
Version: \tab 0.0\cr
Date: \tab 28-March-2010\cr
License: \tab GPL 3\cr
LazyLoad: \tab yes\cr
}
}

\author{
Robert J. Hijmans, Jane Elith and Steven Phillips.
}

\keyword{ package }
\keyword{ spatial }

